Expose mod environment metadata
