+++ src/main/java/net/fabricmc/loader/metadata/BuiltinModMetadata.java
@@ -25,18 +25,16 @@ import java.util.NavigableMap;
 import java.util.Optional;
 import java.util.TreeMap;
 
+import net.fabricmc.api.Environment;
 import net.fabricmc.loader.api.Version;
-import net.fabricmc.loader.api.metadata.ContactInformation;
-import net.fabricmc.loader.api.metadata.CustomValue;
-import net.fabricmc.loader.api.metadata.ModDependency;
-import net.fabricmc.loader.api.metadata.ModMetadata;
-import net.fabricmc.loader.api.metadata.Person;
+import net.fabricmc.loader.api.metadata.*;
 import net.fabricmc.loader.util.version.VersionDeserializer;
 import net.fabricmc.loader.util.version.VersionParsingException;
 
 public final class BuiltinModMetadata extends AbstractModMetadata {
 	private final String id;
 	private final Version version;
+	private final ModEnvironment environment;
 	private final String name;
 	private final String description;
 	private final Collection<Person> authors;
@@ -46,6 +44,7 @@ public final class BuiltinModMetadata extends AbstractModMetadata {
 	private final NavigableMap<Integer, String> icons;
 
 	private BuiltinModMetadata(String id, Version version,
+			ModEnvironment environment,
 			String name, String description,
 			Collection<Person> authors, Collection<Person> contributors,
 			ContactInformation contact,
@@ -53,6 +52,7 @@ public final class BuiltinModMetadata extends AbstractModMetadata {
 			NavigableMap<Integer, String> icons) {
 		this.id = id;
 		this.version = version;
+		this.environment = environment;
 		this.name = name;
 		this.description = description;
 		this.authors = authors;
@@ -75,6 +75,11 @@ public final class BuiltinModMetadata extends AbstractModMetadata {
 	@Override
 	public Version getVersion() {
 		return version;
+	}
+
+	@Override
+	public ModEnvironment getEnvironment() {
+		return environment;
 	}
 
 	@Override
@@ -136,6 +141,7 @@ public final class BuiltinModMetadata extends AbstractModMetadata {
 	public static class Builder {
 		private final String id;
 		private final Version version;
+		private ModEnvironment environment = ModEnvironment.UNIVERSAL;
 		private String name;
 		private String description = "";
 		private final Collection<Person> authors = new ArrayList<>();
@@ -152,6 +158,11 @@ public final class BuiltinModMetadata extends AbstractModMetadata {
 			} catch (VersionParsingException e) {
 				throw new RuntimeException(e);
 			}
+		}
+
+		public Builder setEnvironment(ModEnvironment environment) {
+			this.environment = environment;
+			return this;
 		}
 
 		public Builder setName(String name) {
@@ -190,7 +201,7 @@ public final class BuiltinModMetadata extends AbstractModMetadata {
 		}
 
 		public ModMetadata build() {
-			return new BuiltinModMetadata(id, version, name, description, authors, contributors, contact, license, icons);
+			return new BuiltinModMetadata(id, version, environment, name, description, authors, contributors, contact, license, icons);
 		}
 
 		private static Person createPerson(String name, Map<String, String> contactMap) {
