+++ src/main/java/net/fabricmc/loader/metadata/ModMetadataV0.java
@@ -21,6 +21,7 @@ import net.fabricmc.api.EnvType;
 import net.fabricmc.loader.api.metadata.ContactInformation;
 import net.fabricmc.loader.api.metadata.CustomValue;
 import net.fabricmc.loader.api.metadata.ModDependency;
+import net.fabricmc.loader.api.metadata.ModEnvironment;
 import net.fabricmc.loader.api.Version;
 import org.apache.logging.log4j.Logger;
 
@@ -225,6 +226,18 @@ public class ModMetadataV0 extends AbstractModMetadata implements LoaderModMetad
 	@Override
 	public Collection<ModDependency> getBreaks() {
 		return conflicts.toModDependencies();
+	}
+
+	@Override
+	public ModEnvironment getEnvironment() {
+		switch (side) {
+			case CLIENT:
+				return ModEnvironment.CLIENT;
+			case SERVER:
+				return ModEnvironment.SERVER;
+			default:
+				return ModEnvironment.UNIVERSAL;
+		}
 	}
 
 	public static class Mixins {
